endpoint: provide k8s metadata fetcher via cell
In preparation for extracting the Endpoint API, this commit extracts
the `EndpointMetadataFetcher` into its own Hive Cell `endpoint-metadata`
